+++ py-rattler/src/networking/client.rs
@@ -9,6 +9,8 @@ use rattler_networking::{
 };
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use reqwest_middleware::ClientWithMiddleware;
+use reqwest_retry::policies::ExponentialBackoff;
+use reqwest_retry::RetryTransientMiddleware;
 use std::collections::HashMap;
 
 static RATTLER_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"));
@@ -23,14 +25,19 @@ pub struct PyClientWithMiddleware {
 #[pymethods]
 impl PyClientWithMiddleware {
     #[new]
-    #[pyo3(signature = (middlewares=None, headers=None))]
+    #[pyo3(signature = (middlewares=None, headers=None, timeout=None))]
     pub fn new(
         middlewares: Option<Vec<PyMiddleware>>,
         headers: Option<HashMap<String, String>>,
+        timeout: Option<u64>,
     ) -> PyResult<Self> {
         let middlewares = middlewares.unwrap_or_default();
 
         let mut client_builder = reqwest::Client::builder();
+
+        if let Some(timeout) = timeout {
+            client_builder = client_builder.timeout(std::time::Duration::from_secs(timeout));
+        }
 
         if let Some(headers) = headers {
             let mut header_map = HeaderMap::new();
@@ -45,7 +52,8 @@ impl PyClientWithMiddleware {
             client_builder = client_builder.user_agent(RATTLER_USER_AGENT);
         }
 
-        let mut client = reqwest_middleware::ClientBuilder::new(client_builder.build().unwrap());
+        let reqwest_client = client_builder.build().unwrap();
+        let mut client = reqwest_middleware::ClientBuilder::new(reqwest_client.clone());
 
         for middleware in middlewares {
             match middleware {
@@ -58,8 +66,13 @@ impl PyClientWithMiddleware {
                             .map_err(PyRattlerError::from)?,
                     );
                 }
-                PyMiddleware::Oci(middleware) => {
-                    client = client.with(OciMiddleware::from(middleware));
+                PyMiddleware::Retry(middleware) => {
+                    let policy = ExponentialBackoff::builder()
+                        .build_with_max_retries(middleware.max_retries);
+                    client = client.with(RetryTransientMiddleware::new_with_policy(policy));
+                }
+                PyMiddleware::Oci(_middleware) => {
+                    client = client.with(OciMiddleware::new(reqwest_client.clone()));
                 }
                 PyMiddleware::Gcs(middleware) => {
                     client = client.with(GCSMiddleware::from(middleware));