+++ crates/rattler_index/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.27.16](https://github.com/conda/rattler/compare/rattler_index-v0.27.15...rattler_index-v0.27.16) - 2026-02-25
+
+### Other
+
+- experimental repodata v3 ([#2093](https://github.com/conda/rattler/pull/2093))
+
 ## [0.27.15](https://github.com/conda/rattler/compare/rattler_index-v0.27.14...rattler_index-v0.27.15) - 2026-02-20
 
 ### Other