+++ crates/rattler_lock/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.16](https://github.com/conda/rattler/compare/rattler_lock-v0.26.15...rattler_lock-v0.26.16) - 2026-02-25
+
+### Other
+
+- updated the following local packages: rattler_conda_types, rattler_solve
+
 ## [0.26.15](https://github.com/conda/rattler/compare/rattler_lock-v0.26.14...rattler_lock-v0.26.15) - 2026-02-20
 
 ### Other