+++ crates/rattler_networking/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.2](https://github.com/conda/rattler/compare/rattler_networking-v0.26.1...rattler_networking-v0.26.2) - 2026-02-25
+
+### Fixed
+
+- reuse reqwest client in OCI middleware ([#2089](https://github.com/conda/rattler/pull/2089))
+
 ## [0.26.1](https://github.com/conda/rattler/compare/rattler_networking-v0.26.0...rattler_networking-v0.26.1) - 2026-02-20
 
 ### Other