+++ crates/rattler_s3/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.25](https://github.com/conda/rattler/compare/rattler_s3-v0.1.24...rattler_s3-v0.1.25) - 2026-02-25
+
+### Other
+
+- updated the following local packages: rattler_networking
+
 ## [0.1.24](https://github.com/conda/rattler/compare/rattler_s3-v0.1.23...rattler_s3-v0.1.24) - 2026-02-20
 
 ### Other