+++ crates/rattler_solve/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.6](https://github.com/conda/rattler/compare/rattler_solve-v4.2.5...rattler_solve-v4.2.6) - 2026-02-25
+
+### Other
+
+- updated the following local packages: rattler_conda_types
+
 ## [4.2.5](https://github.com/conda/rattler/compare/rattler_solve-v4.2.4...rattler_solve-v4.2.5) - 2026-02-20
 
 ### Other